backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Topiltzin Hernández Mares <topiltzin.hernandez@crabti.com>"]

[tool.poetry.dependencies]
python = "3.8.1"
django = "^3.2.7"
mysqlclient = "^2.0.3"
flake8 = "^3.9.2"
djangorestframework = "^3.12.4"
Markdown = "^3.3.4"
django-filter = "^2.4.0"
factory-boy = "^3.2.0"
coverage = "^5.5"
django-nose = "^1.4.7"
Faker = "^8.13.2"
django-cors-headers = "^3.8.0"
django-rest-multiple-models = "^2.1.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
